+++ src/aui-button/js/aui-button-core.js
@@ -165,6 +165,8 @@ ButtonExt.prototype = {
         var instance = this;
 
         instance._uiSetIcon(instance.get('icon'));
+
+        instance._setButtonRole();
     },
 
     /**
@@ -191,6 +193,18 @@ ButtonExt.prototype = {
         var instance = this;
 
         instance._uiSetIconAlign(event.newVal);
+    },
+
+    /**
+     * Sets the role attribute on the bounding box to 'button';
+     *
+     * @method setButtonRole
+     */
+    _setButtonRole: function() {
+        var instance = this;
+        var boundingBox = instance.get('boundingBox');
+
+        boundingBox.setAttribute('role', 'button');
     },
 
     /**
@@ -383,6 +397,16 @@ A.mix(ButtonGroup.prototype, {
     CONTENT_TEMPLATE: null,
 
     /**
+     * Constructor for `A.ButtonGroup`. Lifecycle.
+     *
+     * @method initializer
+     * @protected
+     */
+    initializer: function() {
+        this.after('selectionChange', this._afterSelectionChange);
+    },
+
+    /**
      * Returns the `item` or `node` of specified `index`.
      *
      * @method item
@@ -408,7 +432,9 @@ A.mix(ButtonGroup.prototype, {
      * @protected
      */
     renderUI: function() {
-        var instance = this;
+        var instance = this,
+            boundingBox = instance.get('boundingBox'),
+            type = instance.get('type');
 
         instance.getButtons().each(function(button) {
             if (!button.button && !A.instanceOf(A.Widget.getByNode(button), A.Button)) {
@@ -416,6 +442,7 @@ A.mix(ButtonGroup.prototype, {
                 // A.Plugin.Button doesn't current allow augmentation, therefore
                 // it can't add A.ButtonExt extra attributes to it.
                 button.addClass(A.ButtonCore.CLASS_NAMES.BUTTON_DEFAULT);
+                button.setAttribute('role', 'option');
 
                 if (A.Button.hasWidgetLazyConstructorData(button)) {
                     new A.Button(A.Button.getWidgetLazyConstructorFromNodeData(button));
@@ -426,6 +453,13 @@ A.mix(ButtonGroup.prototype, {
                 }
             }
         });
+
+        boundingBox.setAttrs({
+            'aria-multiselectable': (type === 'checkbox') ? true : false,
+            role: 'listbox'
+        });
+
+        instance.syncAriaSelected(instance.getButtons());
     },
 
     /**
@@ -438,6 +472,22 @@ A.mix(ButtonGroup.prototype, {
         var instance = this;
 
         return instance.toggleSelect(items, true);
+    },
+
+    /**
+     * Updates the 'aria-selected' attribute on all buttons.
+     *
+     * @method syncAriaSelected
+     * @param {Array} buttons
+     */
+    syncAriaSelected: function(buttons) {
+        var selected;
+
+        buttons.each(function(button) {
+            selected = button.hasClass(A.ButtonGroup.CLASS_NAMES.SELECTED);
+
+            button.setAttribute('aria-selected', selected);
+        });
     },
 
     /**
@@ -497,5 +547,17 @@ A.mix(ButtonGroup.prototype, {
         var instance = this;
 
         return instance.toggleSelect(items, false);
+    },
+
+    /**
+     * Fires after 'selectionChange' event.
+     *
+     * @method _afterSelectionChange
+     * @protected
+     */
+    _afterSelectionChange: function() {
+        var instance = this;
+
+        instance.syncAriaSelected(instance.getButtons());
     }
 }, true);